+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Improved
 - Added a option to choose how much orders show in HRX Orders list
 - Disabled mass action checkbox for orders, which have error
+- Added Woocommerce Order preview ability in HRX delivery page
 
 ## [1.0.1] - 2022-12-09
 ### Improved